refactor(client): decompose large page components into focused dialog modules (MGG-211)
Extract CRUD dialog logic from 9 page components into 14 dedicated dialog
components, reducing page sizes by 23% overall (AdminUsers -67%, ApiKeys -49%).
Fix undefined function references in ReceiptItems and Transactions pages.
Move navigationMenuTriggerStyle to separate module to resolve ESLint
react-refresh/only-export-components violation.
