+++ src/client/src/components/ResetPasswordDialog.tsx
@@ -1,0 +1,106 @@
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { z } from "zod";
+import { useResetUserPassword } from "@/hooks/useUsers";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Spinner } from "@/components/ui/spinner";
+import {
+  Dialog,
+  DialogContent,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+  DialogDescription,
+} from "@/components/ui/dialog";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+
+const resetPasswordSchema = z.object({
+  newPassword: z.string().min(8, "Password must be at least 8 characters"),
+});
+
+type ResetPasswordFormValues = z.infer<typeof resetPasswordSchema>;
+
+interface ResetPasswordDialogProps {
+  userId: string | null;
+  onClose: () => void;
+}
+
+export function ResetPasswordDialog({
+  userId,
+  onClose,
+}: ResetPasswordDialogProps) {
+  const resetPassword = useResetUserPassword();
+
+  const form = useForm<ResetPasswordFormValues>({
+    resolver: zodResolver(resetPasswordSchema),
+    defaultValues: { newPassword: "" },
+  });
+
+  async function onSubmit(values: ResetPasswordFormValues) {
+    if (!userId) return;
+    await resetPassword.mutateAsync({
+      userId,
+      newPassword: values.newPassword,
+    });
+    onClose();
+    form.reset();
+  }
+
+  function handleOpenChange(open: boolean) {
+    if (!open) {
+      onClose();
+      form.reset();
+    }
+  }
+
+  return (
+    <Dialog open={!!userId} onOpenChange={handleOpenChange}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>Reset Password</DialogTitle>
+          <DialogDescription>
+            Set a new password for this user. They will be required to change it
+            on their next login.
+          </DialogDescription>
+        </DialogHeader>
+        <Form {...form}>
+          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
+            <FormField
+              control={form.control}
+              name="newPassword"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>New Password</FormLabel>
+                  <FormControl>
+                    <Input
+                      type="password"
+                      placeholder="At least 8 characters"
+                      {...field}
+                    />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            <DialogFooter>
+              <Button type="submit" disabled={form.formState.isSubmitting}>
+                {form.formState.isSubmitting && <Spinner size="sm" />}
+                {form.formState.isSubmitting
+                  ? "Resetting..."
+                  : "Reset Password"}
+              </Button>
+            </DialogFooter>
+          </form>
+        </Form>
+      </DialogContent>
+    </Dialog>
+  );
+}
