+++ src/client/src/pages/Accounts.tsx
@@ -1,10 +1,5 @@
 import { useState, useMemo, useEffect } from "react";
-import {
-  useAccounts,
-  useCreateAccount,
-  useUpdateAccount,
-  useDeleteAccounts,
-} from "@/hooks/useAccounts";
+import { useAccounts } from "@/hooks/useAccounts";
 import { usePageTitle } from "@/hooks/usePageTitle";
 import { useFuzzySearch } from "@/hooks/useFuzzySearch";
 import { useSavedFilters } from "@/hooks/useSavedFilters";
@@ -13,7 +8,7 @@ import { useListKeyboardNav } from "@/hooks/useListKeyboardNav";
 import type { FuseSearchConfig, FilterDefinition } from "@/lib/search";
 import { applyFilters } from "@/lib/search";
 import type { FilterValues } from "@/components/FilterPanel";
-import { AccountForm } from "@/components/AccountForm";
+import { AccountDialogs } from "@/components/AccountDialogs";
 import { FuzzySearchInput } from "@/components/FuzzySearchInput";
 import { FilterPanel } from "@/components/FilterPanel";
 import type { FilterField } from "@/components/FilterPanel";
@@ -24,12 +19,6 @@ import { Pagination } from "@/components/Pagination";
 import { Button } from "@/components/ui/button";
 import { Badge } from "@/components/ui/badge";
 import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import {
   Table,
   TableBody,
   TableCell,
@@ -38,7 +27,6 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import { TableSkeleton } from "@/components/ui/table-skeleton";
-import { Spinner } from "@/components/ui/spinner";
 
 interface AccountResponse {
   id: string;
@@ -65,9 +53,6 @@ const FILTER_DEFS: FilterDefinition[] = [
 function Accounts() {
   usePageTitle("Accounts");
   const { data: accounts, isLoading } = useAccounts();
-  const createAccount = useCreateAccount();
-  const updateAccount = useUpdateAccount();
-  const deleteAccounts = useDeleteAccounts();
 
   const [selected, setSelected] = useState<Set<string>>(new Set());
   const [createOpen, setCreateOpen] = useState(false);
@@ -301,85 +286,19 @@ function Accounts() {
         </>
       )}
 
-      {/* Create Dialog */}
-      <Dialog open={createOpen} onOpenChange={setCreateOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Create Account</DialogTitle>
-          </DialogHeader>
-          <AccountForm
-            mode="create"
-            isSubmitting={createAccount.isPending}
-            onCancel={() => setCreateOpen(false)}
-            onSubmit={(values) => {
-              createAccount.mutate(values, {
-                onSuccess: () => setCreateOpen(false),
-              });
-            }}
-          />
-        </DialogContent>
-      </Dialog>
-
-      {/* Edit Dialog */}
-      <Dialog
-        open={editAccount !== null}
-        onOpenChange={(open) => !open && setEditAccount(null)}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Edit Account</DialogTitle>
-          </DialogHeader>
-          {editAccount && (
-            <AccountForm
-              mode="edit"
-              defaultValues={{
-                accountCode: editAccount.accountCode,
-                name: editAccount.name,
-                isActive: editAccount.isActive,
-              }}
-              isSubmitting={updateAccount.isPending}
-              onCancel={() => setEditAccount(null)}
-              onSubmit={(values) => {
-                updateAccount.mutate(
-                  { id: editAccount.id, ...values },
-                  { onSuccess: () => setEditAccount(null) },
-                );
-              }}
-            />
-          )}
-        </DialogContent>
-      </Dialog>
-
-      {/* Delete Confirmation Dialog */}
-      <Dialog open={deleteOpen} onOpenChange={setDeleteOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Delete Accounts</DialogTitle>
-          </DialogHeader>
-          <p className="text-sm text-muted-foreground">
-            Are you sure you want to delete {selected.size} account(s)? This
-            action can be undone by restoring.
-          </p>
-          <div className="flex justify-end gap-2 pt-4">
-            <Button variant="outline" onClick={() => setDeleteOpen(false)}>
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              disabled={deleteAccounts.isPending}
-              onClick={() => {
-                const ids = [...selected];
-                setSelected(new Set());
-                setDeleteOpen(false);
-                deleteAccounts.mutate(ids);
-              }}
-            >
-              {deleteAccounts.isPending && <Spinner size="sm" />}
-              {deleteAccounts.isPending ? "Deleting..." : "Delete"}
-            </Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <AccountDialogs
+        createOpen={createOpen}
+        onCreateOpenChange={setCreateOpen}
+        editAccount={editAccount}
+        onEditClose={() => setEditAccount(null)}
+        deleteOpen={deleteOpen}
+        onDeleteOpenChange={setDeleteOpen}
+        selectedIds={[...selected]}
+        onDeleteComplete={() => {
+          setSelected(new Set());
+          setDeleteOpen(false);
+        }}
+      />
     </div>
   );
 }
