+++ src/client/src/pages/AdminUsers.tsx
@@ -1,21 +1,15 @@
 import { useState } from "react";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
-import {
-  useUsers,
-  useCreateUser,
-  useUpdateUser,
-  useDeleteUser,
-  useResetUserPassword,
-} from "@/hooks/useUsers";
+import { useUsers } from "@/hooks/useUsers";
 import { useAuth } from "@/hooks/useAuth";
 import { usePageTitle } from "@/hooks/usePageTitle";
 import { useListKeyboardNav } from "@/hooks/useListKeyboardNav";
+import { CreateUserDialog } from "@/components/CreateUserDialog";
+import { EditUserDialog } from "@/components/EditUserDialog";
+import type { EditableUser } from "@/components/EditUserDialog";
+import { ResetPasswordDialog } from "@/components/ResetPasswordDialog";
+import { DeactivateUserDialog } from "@/components/DeactivateUserDialog";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
 import { Badge } from "@/components/ui/badge";
-import { Spinner } from "@/components/ui/spinner";
 import {
   Table,
   TableBody,
@@ -25,68 +19,11 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import { TableSkeleton } from "@/components/ui/table-skeleton";
-import {
-  Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogDescription,
-} from "@/components/ui/dialog";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Label } from "@/components/ui/label";
 
-const AVAILABLE_ROLES = ["Admin", "User"];
-
-const createUserSchema = z.object({
-  email: z.string().email("Please enter a valid email address"),
-  password: z.string().min(8, "Password must be at least 8 characters"),
-  role: z.string().min(1, "Role is required"),
-  firstName: z.string().optional(),
-  lastName: z.string().optional(),
-});
-
-type CreateUserFormValues = z.infer<typeof createUserSchema>;
-
-const editUserSchema = z.object({
-  email: z.string().email("Please enter a valid email address"),
-  role: z.string().min(1, "Role is required"),
-  isDisabled: z.boolean(),
-  firstName: z.string().optional(),
-  lastName: z.string().optional(),
-});
-
-type EditUserFormValues = z.infer<typeof editUserSchema>;
-
-const resetPasswordSchema = z.object({
-  newPassword: z.string().min(8, "Password must be at least 8 characters"),
-});
-
-type ResetPasswordFormValues = z.infer<typeof resetPasswordSchema>;
+function formatDate(dateStr: string | null | undefined) {
+  if (!dateStr) return "-";
+  return new Date(dateStr).toLocaleDateString();
+}
 
 function AdminUsers() {
   usePageTitle("User Management");
@@ -96,20 +33,8 @@ function AdminUsers() {
   const pageSize = 20;
   const { data, isLoading } = useUsers(page, pageSize);
 
-  const createUser = useCreateUser();
-  const updateUser = useUpdateUser();
-  const deleteUser = useDeleteUser();
-  const resetPassword = useResetUserPassword();
-
   const [createOpen, setCreateOpen] = useState(false);
-  const [editUser, setEditUser] = useState<{
-    id: string;
-    email: string;
-    firstName?: string | null;
-    lastName?: string | null;
-    role: string;
-    isDisabled: boolean;
-  } | null>(null);
+  const [editUser, setEditUser] = useState<EditableUser | null>(null);
   const [resetUserId, setResetUserId] = useState<string | null>(null);
   const [deactivateUser, setDeactivateUser] = useState<{
     id: string;
@@ -127,63 +52,6 @@ function AdminUsers() {
     enabled: listItems.length > 0,
   });
 
-  const createForm = useForm<CreateUserFormValues>({
-    resolver: zodResolver(createUserSchema),
-    defaultValues: {
-      email: "",
-      password: "",
-      role: "User",
-      firstName: "",
-      lastName: "",
-    },
-  });
-
-  const editForm = useForm<EditUserFormValues>({
-    resolver: zodResolver(editUserSchema),
-  });
-
-  const resetForm = useForm<ResetPasswordFormValues>({
-    resolver: zodResolver(resetPasswordSchema),
-    defaultValues: { newPassword: "" },
-  });
-
-  async function onCreateSubmit(values: CreateUserFormValues) {
-    await createUser.mutateAsync({
-      email: values.email,
-      password: values.password,
-      role: values.role,
-      firstName: values.firstName || null,
-      lastName: values.lastName || null,
-    });
-    setCreateOpen(false);
-    createForm.reset();
-  }
-
-  async function onEditSubmit(values: EditUserFormValues) {
-    if (!editUser) return;
-    await updateUser.mutateAsync({
-      userId: editUser.id,
-      body: {
-        email: values.email,
-        role: values.role,
-        isDisabled: values.isDisabled,
-        firstName: values.firstName || null,
-        lastName: values.lastName || null,
-      },
-    });
-    setEditUser(null);
-  }
-
-  async function onResetSubmit(values: ResetPasswordFormValues) {
-    if (!resetUserId) return;
-    await resetPassword.mutateAsync({
-      userId: resetUserId,
-      newPassword: values.newPassword,
-    });
-    setResetUserId(null);
-    resetForm.reset();
-  }
-
   function openEdit(user: (typeof items)[0]) {
     const primaryRole = user.roles[0] ?? "User";
     setEditUser({
@@ -194,18 +62,6 @@ function AdminUsers() {
       role: primaryRole,
       isDisabled: user.isDisabled,
     });
-    editForm.reset({
-      email: user.email,
-      role: primaryRole,
-      isDisabled: user.isDisabled,
-      firstName: user.firstName ?? "",
-      lastName: user.lastName ?? "",
-    });
-  }
-
-  function formatDate(dateStr: string | null | undefined) {
-    if (!dateStr) return "-";
-    return new Date(dateStr).toLocaleDateString();
   }
 
   return (
@@ -349,330 +205,16 @@ function AdminUsers() {
         </>
       )}
 
-      {/* Create User Dialog */}
-      <Dialog open={createOpen} onOpenChange={setCreateOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Create User</DialogTitle>
-            <DialogDescription>
-              Create a new user account. They will be required to change their
-              password on first login.
-            </DialogDescription>
-          </DialogHeader>
-          <Form {...createForm}>
-            <form
-              onSubmit={createForm.handleSubmit(onCreateSubmit)}
-              className="space-y-4"
-            >
-              <div className="grid grid-cols-2 gap-4">
-                <FormField
-                  control={createForm.control}
-                  name="firstName"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>First Name</FormLabel>
-                      <FormControl>
-                        <Input placeholder="John" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={createForm.control}
-                  name="lastName"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Last Name</FormLabel>
-                      <FormControl>
-                        <Input placeholder="Doe" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-              <FormField
-                control={createForm.control}
-                name="email"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Email</FormLabel>
-                    <FormControl>
-                      <Input
-                        type="email"
-                        placeholder="name@example.com"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={createForm.control}
-                name="password"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Password</FormLabel>
-                    <FormControl>
-                      <Input
-                        type="password"
-                        placeholder="At least 8 characters"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={createForm.control}
-                name="role"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Role</FormLabel>
-                    <Select
-                      onValueChange={field.onChange}
-                      defaultValue={field.value}
-                    >
-                      <FormControl>
-                        <SelectTrigger className="w-full">
-                          <SelectValue placeholder="Select a role" />
-                        </SelectTrigger>
-                      </FormControl>
-                      <SelectContent>
-                        {AVAILABLE_ROLES.map((role) => (
-                          <SelectItem key={role} value={role}>
-                            {role}
-                          </SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <DialogFooter>
-                <Button
-                  type="submit"
-                  disabled={createForm.formState.isSubmitting}
-                >
-                  {createForm.formState.isSubmitting && <Spinner size="sm" />}
-                  {createForm.formState.isSubmitting
-                    ? "Creating..."
-                    : "Create User"}
-                </Button>
-              </DialogFooter>
-            </form>
-          </Form>
-        </DialogContent>
-      </Dialog>
-
-      {/* Edit User Dialog */}
-      <Dialog open={!!editUser} onOpenChange={() => setEditUser(null)}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Edit User</DialogTitle>
-            <DialogDescription>
-              Update user details, role, and account status.
-            </DialogDescription>
-          </DialogHeader>
-          <Form {...editForm}>
-            <form
-              onSubmit={editForm.handleSubmit(onEditSubmit)}
-              className="space-y-4"
-            >
-              <div className="grid grid-cols-2 gap-4">
-                <FormField
-                  control={editForm.control}
-                  name="firstName"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>First Name</FormLabel>
-                      <FormControl>
-                        <Input placeholder="John" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={editForm.control}
-                  name="lastName"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Last Name</FormLabel>
-                      <FormControl>
-                        <Input placeholder="Doe" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-              <FormField
-                control={editForm.control}
-                name="email"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Email</FormLabel>
-                    <FormControl>
-                      <Input type="email" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={editForm.control}
-                name="role"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Role</FormLabel>
-                    <Select
-                      onValueChange={field.onChange}
-                      value={field.value}
-                    >
-                      <FormControl>
-                        <SelectTrigger className="w-full">
-                          <SelectValue />
-                        </SelectTrigger>
-                      </FormControl>
-                      <SelectContent>
-                        {AVAILABLE_ROLES.map((role) => (
-                          <SelectItem key={role} value={role}>
-                            {role}
-                          </SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={editForm.control}
-                name="isDisabled"
-                render={({ field }) => (
-                  <FormItem>
-                    <div className="flex items-center gap-3">
-                      <FormControl>
-                        <input
-                          type="checkbox"
-                          checked={field.value}
-                          onChange={field.onChange}
-                          className="h-4 w-4 rounded border-gray-300"
-                          aria-label="Disable account"
-                        />
-                      </FormControl>
-                      <Label>Account Disabled</Label>
-                    </div>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <DialogFooter>
-                <Button
-                  type="submit"
-                  disabled={editForm.formState.isSubmitting}
-                >
-                  {editForm.formState.isSubmitting && <Spinner size="sm" />}
-                  {editForm.formState.isSubmitting
-                    ? "Saving..."
-                    : "Save Changes"}
-                </Button>
-              </DialogFooter>
-            </form>
-          </Form>
-        </DialogContent>
-      </Dialog>
-
-      {/* Reset Password Dialog */}
-      <Dialog
-        open={!!resetUserId}
-        onOpenChange={() => {
-          setResetUserId(null);
-          resetForm.reset();
-        }}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Reset Password</DialogTitle>
-            <DialogDescription>
-              Set a new password for this user. They will be required to change
-              it on their next login.
-            </DialogDescription>
-          </DialogHeader>
-          <Form {...resetForm}>
-            <form
-              onSubmit={resetForm.handleSubmit(onResetSubmit)}
-              className="space-y-4"
-            >
-              <FormField
-                control={resetForm.control}
-                name="newPassword"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>New Password</FormLabel>
-                    <FormControl>
-                      <Input
-                        type="password"
-                        placeholder="At least 8 characters"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <DialogFooter>
-                <Button
-                  type="submit"
-                  disabled={resetForm.formState.isSubmitting}
-                >
-                  {resetForm.formState.isSubmitting && <Spinner size="sm" />}
-                  {resetForm.formState.isSubmitting
-                    ? "Resetting..."
-                    : "Reset Password"}
-                </Button>
-              </DialogFooter>
-            </form>
-          </Form>
-        </DialogContent>
-      </Dialog>
-
-      {/* Deactivate Confirmation */}
-      <AlertDialog
-        open={!!deactivateUser}
-        onOpenChange={() => setDeactivateUser(null)}
-      >
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Deactivate User</AlertDialogTitle>
-            <AlertDialogDescription>
-              Are you sure you want to deactivate{" "}
-              <span className="font-semibold">{deactivateUser?.email}</span>?
-              They will no longer be able to log in.
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel>Cancel</AlertDialogCancel>
-            <AlertDialogAction
-              variant="destructive"
-              onClick={async () => {
-                if (deactivateUser) {
-                  await deleteUser.mutateAsync(deactivateUser.id);
-                  setDeactivateUser(null);
-                }
-              }}
-            >
-              {deleteUser.isPending && <Spinner size="sm" />}
-              Deactivate
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
+      <CreateUserDialog open={createOpen} onOpenChange={setCreateOpen} />
+      <EditUserDialog user={editUser} onClose={() => setEditUser(null)} />
+      <ResetPasswordDialog
+        userId={resetUserId}
+        onClose={() => setResetUserId(null)}
+      />
+      <DeactivateUserDialog
+        user={deactivateUser}
+        onClose={() => setDeactivateUser(null)}
+      />
     </div>
   );
 }
