+++ src/client/src/pages/ApiKeys.tsx
@@ -1,12 +1,13 @@
 import { useState, useEffect } from "react";
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { usePageTitle } from "@/hooks/usePageTitle";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
+import { useListKeyboardNav } from "@/hooks/useListKeyboardNav";
 import client from "@/lib/api-client";
-import { showSuccess, showError } from "@/lib/toast";
+import { CreateApiKeyDialog } from "@/components/CreateApiKeyDialog";
+import { CreatedKeyDialog } from "@/components/CreatedKeyDialog";
+import { RevokeApiKeyDialog } from "@/components/RevokeApiKeyDialog";
 import { Button } from "@/components/ui/button";
+import { Badge } from "@/components/ui/badge";
 import {
   Card,
   CardContent,
@@ -14,25 +15,6 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Alert, AlertDescription } from "@/components/ui/alert";
-import { Badge } from "@/components/ui/badge";
-import { useListKeyboardNav } from "@/hooks/useListKeyboardNav";
 import {
   Table,
   TableBody,
@@ -42,14 +24,6 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import { TableSkeleton } from "@/components/ui/table-skeleton";
-import { Spinner } from "@/components/ui/spinner";
-
-const createKeySchema = z.object({
-  name: z.string().min(1, "Name is required"),
-  expiresAt: z.string().optional(),
-});
-
-type CreateKeyFormValues = z.infer<typeof createKeySchema>;
 
 function getKeyStatus(key: {
   isRevoked: boolean;
@@ -80,7 +54,6 @@ function formatDate(dateStr: string | null | undefined): string {
 
 function ApiKeys() {
   usePageTitle("API Keys");
-  const queryClient = useQueryClient();
   const [createOpen, setCreateOpen] = useState(false);
   const [createdKey, setCreatedKey] = useState<string | null>(null);
   const [revokeId, setRevokeId] = useState<string | null>(null);
@@ -96,77 +69,13 @@ function ApiKeys() {
     },
   });
 
-  const createMutation = useMutation({
-    mutationFn: async (values: CreateKeyFormValues) => {
-      const { data, error } = await client.POST("/api/apikeys", {
-        body: {
-          name: values.name,
-          expiresAt: values.expiresAt || undefined,
-        },
-      });
-      if (error) throw error;
-      return data;
-    },
-    onSuccess: (data) => {
-      queryClient.invalidateQueries({ queryKey: ["apiKeys"] });
-      if (data) {
-        setCreatedKey(data.rawKey);
-      }
-      setCreateOpen(false);
-    },
-    onError: () => {
-      showError("Failed to create API key.");
-    },
-  });
-
-  const revokeMutation = useMutation({
-    mutationFn: async (id: string) => {
-      const { error } = await client.DELETE("/api/apikeys/{id}", {
-        params: { path: { id } },
-      });
-      if (error) throw error;
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["apiKeys"] });
-      showSuccess("API key revoked.");
-      setRevokeId(null);
-    },
-    onError: () => {
-      showError("Failed to revoke API key.");
-    },
-  });
-
-  const createForm = useForm<CreateKeyFormValues>({
-    resolver: zodResolver(createKeySchema),
-    defaultValues: { name: "", expiresAt: "" },
-  });
-
-  function handleCreateOpen() {
-    createForm.reset();
-    setCreateOpen(true);
-  }
-
   useEffect(() => {
     function onNewItem() {
-      handleCreateOpen();
+      setCreateOpen(true);
     }
     window.addEventListener("shortcut:new-item", onNewItem);
     return () => window.removeEventListener("shortcut:new-item", onNewItem);
   });
-
-  function handleCreateSubmit(values: CreateKeyFormValues) {
-    createMutation.mutate(values);
-  }
-
-  async function handleCopyKey() {
-    if (!createdKey) return;
-    try {
-      await navigator.clipboard.writeText(createdKey);
-      showSuccess("API key copied to clipboard.");
-    } catch {
-      showError("Failed to copy to clipboard.");
-    }
-  }
 
   const { focusedId, setFocusedIndex, tableRef } = useListKeyboardNav({
     items: apiKeys as { id: string }[],
@@ -183,7 +92,7 @@ function ApiKeys() {
             Manage API keys for programmatic access
           </p>
         </div>
-        <Button onClick={handleCreateOpen}>Create API Key</Button>
+        <Button onClick={() => setCreateOpen(true)}>Create API Key</Button>
       </div>
 
       <Card>
@@ -203,175 +112,79 @@ function ApiKeys() {
             </p>
           ) : (
             <div ref={tableRef}>
-            <Table>
-              <TableHeader>
-                <TableRow>
-                  <TableHead>Name</TableHead>
-                  <TableHead>Created</TableHead>
-                  <TableHead>Last Used</TableHead>
-                  <TableHead>Expires</TableHead>
-                  <TableHead>Status</TableHead>
-                  <TableHead className="text-right">Actions</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {apiKeys.map((key, index) => {
-                  const status = getKeyStatus(key);
-                  return (
-                    <TableRow
-                      key={key.id}
-                      className={`cursor-pointer ${focusedId === key.id ? "bg-accent" : ""}`}
-                      onClick={(e) => {
-                        if ((e.target as HTMLElement).closest("button, input, a, [role='button']")) return;
-                        setFocusedIndex(index);
-                      }}
-                    >
-                      <TableCell className="font-medium">{key.name}</TableCell>
-                      <TableCell>{formatDate(key.createdAt)}</TableCell>
-                      <TableCell>{formatDate(key.lastUsedAt)}</TableCell>
-                      <TableCell>{formatDate(key.expiresAt)}</TableCell>
-                      <TableCell>
-                        <Badge variant={statusBadgeVariant(status)}>
-                          {status}
-                        </Badge>
-                      </TableCell>
-                      <TableCell className="text-right">
-                        {status === "active" && (
-                          <Button
-                            variant="destructive"
-                            size="sm"
-                            onClick={() => setRevokeId(key.id)}
-                          >
-                            Revoke
-                          </Button>
-                        )}
-                      </TableCell>
-                    </TableRow>
-                  );
-                })}
-              </TableBody>
-            </Table>
+              <Table>
+                <TableHeader>
+                  <TableRow>
+                    <TableHead>Name</TableHead>
+                    <TableHead>Created</TableHead>
+                    <TableHead>Last Used</TableHead>
+                    <TableHead>Expires</TableHead>
+                    <TableHead>Status</TableHead>
+                    <TableHead className="text-right">Actions</TableHead>
+                  </TableRow>
+                </TableHeader>
+                <TableBody>
+                  {apiKeys.map((key, index) => {
+                    const status = getKeyStatus(key);
+                    return (
+                      <TableRow
+                        key={key.id}
+                        className={`cursor-pointer ${focusedId === key.id ? "bg-accent" : ""}`}
+                        onClick={(e) => {
+                          if (
+                            (e.target as HTMLElement).closest(
+                              "button, input, a, [role='button']",
+                            )
+                          )
+                            return;
+                          setFocusedIndex(index);
+                        }}
+                      >
+                        <TableCell className="font-medium">
+                          {key.name}
+                        </TableCell>
+                        <TableCell>{formatDate(key.createdAt)}</TableCell>
+                        <TableCell>{formatDate(key.lastUsedAt)}</TableCell>
+                        <TableCell>{formatDate(key.expiresAt)}</TableCell>
+                        <TableCell>
+                          <Badge variant={statusBadgeVariant(status)}>
+                            {status}
+                          </Badge>
+                        </TableCell>
+                        <TableCell className="text-right">
+                          {status === "active" && (
+                            <Button
+                              variant="destructive"
+                              size="sm"
+                              onClick={() => setRevokeId(key.id)}
+                            >
+                              Revoke
+                            </Button>
+                          )}
+                        </TableCell>
+                      </TableRow>
+                    );
+                  })}
+                </TableBody>
+              </Table>
             </div>
           )}
         </CardContent>
       </Card>
 
-      {/* Create API Key Dialog */}
-      <Dialog open={createOpen} onOpenChange={setCreateOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Create API Key</DialogTitle>
-            <DialogDescription>
-              Generate a new API key for programmatic access.
-            </DialogDescription>
-          </DialogHeader>
-          <Form {...createForm}>
-            <form
-              onSubmit={createForm.handleSubmit(handleCreateSubmit)}
-              className="space-y-4"
-            >
-              <FormField
-                control={createForm.control}
-                name="name"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Name</FormLabel>
-                    <FormControl>
-                      <Input
-                        placeholder="e.g. Paperless Integration"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={createForm.control}
-                name="expiresAt"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Expiration Date (optional)</FormLabel>
-                    <FormControl>
-                      <Input type="date" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <Button
-                type="submit"
-                className="w-full"
-                disabled={createMutation.isPending}
-              >
-                {createMutation.isPending && <Spinner size="sm" />}
-                {createMutation.isPending ? "Creating..." : "Create Key"}
-              </Button>
-            </form>
-          </Form>
-        </DialogContent>
-      </Dialog>
-
-      {/* Created Key Display Dialog */}
-      <Dialog
-        open={createdKey !== null}
-        onOpenChange={(open) => {
-          if (!open) setCreatedKey(null);
-        }}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>API Key Created</DialogTitle>
-          </DialogHeader>
-          <Alert>
-            <AlertDescription>
-              Save this key now. You won&apos;t be able to see it again.
-            </AlertDescription>
-          </Alert>
-          <Input
-            readOnly
-            value={createdKey ?? ""}
-            onClick={(e) => (e.target as HTMLInputElement).select()}
-            className="font-mono text-sm"
-          />
-          <Button onClick={handleCopyKey} className="w-full">
-            Copy to Clipboard
-          </Button>
-        </DialogContent>
-      </Dialog>
-
-      {/* Revoke Confirmation Dialog */}
-      <Dialog
-        open={revokeId !== null}
-        onOpenChange={(open) => {
-          if (!open) setRevokeId(null);
-        }}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Revoke API Key</DialogTitle>
-            <DialogDescription>
-              This action cannot be undone. The API key will be immediately
-              invalidated.
-            </DialogDescription>
-          </DialogHeader>
-          <div className="flex gap-2 justify-end">
-            <Button variant="outline" onClick={() => setRevokeId(null)}>
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              disabled={revokeMutation.isPending}
-              onClick={() => {
-                if (revokeId) revokeMutation.mutate(revokeId);
-              }}
-            >
-              {revokeMutation.isPending && <Spinner size="sm" />}
-              {revokeMutation.isPending ? "Revoking..." : "Revoke"}
-            </Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <CreateApiKeyDialog
+        open={createOpen}
+        onOpenChange={setCreateOpen}
+        onKeyCreated={setCreatedKey}
+      />
+      <CreatedKeyDialog
+        rawKey={createdKey}
+        onClose={() => setCreatedKey(null)}
+      />
+      <RevokeApiKeyDialog
+        keyId={revokeId}
+        onClose={() => setRevokeId(null)}
+      />
     </div>
   );
 }
